在新年时输出一些特别的话，改正“ lost 了当前天数”的语法问题

+++ source/main.h
@@ -14,7 +14,7 @@
 
 /* Structures */
 struct config_struct {
-    uint16_t max_progress_bar_width;
+    uint32_t max_progress_bar_width;
 };
 
 /* Error Type */
